#4 bookscontroller byid, update, delete, create done UNTESTED

+++ backend/src/controllers/book.controller.ts
@@ -3,6 +3,7 @@ const BookModel = require("../models/book.model")
 import { bookSchema } from "../models/book.model"
 import * as tools from "../tools/tools"
 import { Projection } from "../models/projection.model"
+import { UserController } from "./user.controller"
 
 export class BookController{
     
@@ -32,7 +33,6 @@ export class BookController{
             if(genre){
                 const genrePath = BookModel.schema.path("genre")
                 const genres: string[] = genrePath.options.enum.map(function(genre: string){return genre.toLowerCase()})
-                console.log(genres)
                 if(!genres.includes(genre.toLowerCase())){
                     return res.status(400).json({error: "Invalid genre requested"})
                 }
@@ -88,37 +88,72 @@ export class BookController{
     }
 
     static async getBookById(req:any,res:any){
-        const {id} = req.params
         try{
-            const book = await BookModel.findById(id)
-            res.status(200).json(book)
+            const {id, fields} = req.params
+            let allowedFields = ["_id","title","author","user_id","genre","release","description","added_at","updated_at"]
+            
+            const requestedFields: string[] = fields ? fields.split(",") : allowedFields
+            const validFields: string[] = requestedFields.filter(field =>allowedFields.includes(field))
+
+            if(validFields.length === 0) return res.status(400).json({error:"Invalid fields requested"})
+            
+            if(!validFields.includes("_id")) validFields.push("-_id")
+
+            
+            const data = await BookModel.findById(id).select(validFields)
+            if(data){
+                res.status(200).json(data)
+            }
+            else{
+                res.status(404).json({message: "User not found"})
+            }
         }catch(error:any){
             res.status(500).json({message:error.message})
         }
     }
 
     static async createBook(req:any,res:any){
-        const book = req.body
-        const newBook = new BookModel({
-            title: book.title,
-            author: book.author,
-            release: book.release,
-            genre: book.genre,
-            user_id: book.user_id,
-            description: book.description
-        })
         try{
+            let {title, author = "Unknown", genre = "None", description = "", release = null} = req.body
+            
+            if(!title) return res.status(400).json({error: "title required"})
+            
+            if(!UserController.verifyUser(req,"",["editor"])) return res.status(401).json({message:  `Unauthorized`})
+            
+            if(release){
+                if(!tools.isValidISODate(release))
+                    return res.status(400).json({error:"Invalid date"})
+            }
+
+
+            const newBook = new BookModel({
+                title: title,
+                author: author,
+                genre: genre,
+                description: description,
+                release: release,
+                user_id: req.user.id
+            })
             await newBook.save()
             res.status(201).json(newBook)
         }
         catch(error:any){
-            res.status(500).json({message:error.message})
+            UserController.HandleMongooseErrors(error,res)
         }
     }
 
     static async deleteBook(req:any,res:any){
         try{
             const id = req.params
+            if(id){
+                if(!Types.ObjectId.isValid(id)){
+                    return res.status(400).json({message: "Invalid id format"})
+                }
+            }
+            else{
+                return res.status(400).json({message: "id is required"})
+            }
+            if(!UserController.verifyUser(req,"",["editor"])) return res.json(401).send()
             const data = await BookModel.findByIdAndDelete(id)
             if(data){
                 res.status(200).json({message:"Book deleted"})
@@ -134,18 +169,43 @@ export class BookController{
 
     static async updateBook(req:any,res:any){
         try{
-            const id = req.params
-            const book = req.body
-            const updatedBook = await BookModel.findByIdAndUpdate(id,book,{new:true})
-            if(updatedBook){
-                res.status(200).json(updatedBook)
+            const {id} = req.params
+            if(id){
+                if(!Types.ObjectId.isValid(id)){
+                    return res.status(400).json({message: "Invalid id format"})
+                }
             }
             else{
-                res.status(404).json({message:"Book not found"})
+                return res.status(400).json({message: "id is required"})
             }
+            if(!UserController.verifyUser(req,id)) return res.status(401).json({message: "Unauthorized"})
+                const book = await BookModel.findById(id)
+            const updates = req.body
+            const allowedFields = ["title","author","release","genre","description"]
+            if(!book) return res.status(404).json({message: "Book not found"})
+            for (const key of Object.keys(updates)) {
+                if (!allowedFields.includes(key)) {
+                    return res.status(400).json({ message: `Cannot update ${key} field` });
+                }
+                if (key === "release") {
+                    if(!tools.isValidISODate(updates[key])){
+                        return res.status(400).json({message: "Invalid date"})
+                    }
+                } 
+                if (key === "genre") {
+                    const genrePath = BookModel.schema.path("genre")
+                    const genres: string[] = genrePath.options.enum.map(function(genre: string){return genre.toLowerCase()})
+                    if(!genres.includes(updates[key].toLowerCase())){
+                        return res.status(400).json({error: "Invalid genre requested"})
+                    }
+                }
+                book[key] = updates[key];
+            }
+            await book.save()
+            return res.status(200).json({book:book})
         }
         catch(error:any){
-            res.status(500).json({message:error.message})
+            UserController.HandleMongooseErrors(error,res)
         }
     }
 }
